Modify(#86): place 조회할 때 rating 값도 넘겨주기

+++ src/main/java/shop/babsim/babsim/review/domain/Review.java
@@ -26,7 +26,7 @@ import shop.babsim.babsim.place.domain.Place;
 @NoArgsConstructor
 public class Review extends BaseEntity {
 
-    private int rating; // 내가 준 평점
+    private double rating; // 내가 준 평점
 
     private String content;
